Increase number of trials on experimental blocks
The old number was fine, but most subjects took under an hour, which is an
inefficient use of 2 subject-pool points.  This change increases the total
to 448 experimental and 2 practice trials, which leads to 28 trials per
cell.  The number seems a reasonable compromise between getting enough
trials and not allowing slow subjects to run into 3-point territory.

Also advance the minor version number.