Change gap types to different luminance increments.
Replace the stimulus-flash and background-flash conditions with two levels
of luminance increment: a 2:1 change and a 5:1 change.  According to
Rauschenberger (2003), the smaller change should not disrupt attention to
the objects, but the bigger change should.

We will be using two different testing rooms for this experiment with
different monitors, so the code sets different luminance levels separately
for each room.

As part of this change, also simplify the way colors are specified
throughout the code, and advance the minor version number.