Make y-axis limits consistent with those used for ProbeTrack15.
This change makes it easier to compare performance between the two
conditions.

The blank-gap condition in the two experiments are identical.  The only
difference between the experiments is that the point system was explained
to subjects in ProbeTrack15, but not in ProbeTrack13.

It is interesting that RTs in ProbeTrack13 are not reliably affected by
probe delay, whereas they are in ProbeTrack15.  Also, RTs are about 50 ms
faster in ProbeTrack15 than in ProbeTrack13.